Use foreach and cleanup
git-svn-id: https://svn.apache.org/repos/asf/xmlgraphics/batik/trunk@1804130 13f79535-47bb-0310-9956-ffa450edef68

+++ test-sources/org/apache/batik/gvt/TextSelectionTest.java
@@ -126,8 +126,8 @@ public class TextSelectionTest extends JSVGRenderingAccuracyTest {
     public TextSelectionTest(String file,   String textID,
                              Integer start, Integer end) {
         this.textID    = textID;
-        this.start = start.intValue();
-        this.end   = end.intValue();
+        this.start = start;
+        this.end   = end;
         super.setFile(file);
     }
